% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefrcure.R
\name{coefrcure}
\alias{coefrcure}
\title{Retrieves the estimated coefficients from rcure object}
\usage{
coefrcure(x, ...)
}
\arguments{
\item{x}{an object of rcure}

\item{...}{further arguments to be passed to the coefrcure function}
}
\description{
Retrieves the estimated coefficients from rcure object
}
\references{
Cai, C., Zou, Y., Peng, Y., & Zhang, J. (2012). smcure: An R-Package for estimating semiparametric mixture cure models. Computer methods and programs in biomedicine, 108(3), 1255-1260
}

