% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva_cure.R
\name{eva_cure}
\alias{eva_cure}
\title{Evaluate Mixture cure model fitted by rcure}
\usage{
eva_cure(time, delta, X, beta, Z, b, surv, model, baseline = T)
}
\arguments{
\item{time}{survival time}

\item{delta}{event status (1 = event)}

\item{X}{design matrix of survival part}

\item{beta}{parameters of survival part}

\item{Z}{design matrix of cure part}

\item{b}{parameters of cure part}

\item{surv}{survival probability}

\item{model}{the type of survival model ("PH", "PO","Normal")}

\item{baseline}{whether the surv is baseline survival probability}
}
\description{
Evaluate Mixture cure model fitted by rcure
}

