% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva_rcure.R
\name{eva_rcure}
\alias{eva_rcure}
\title{Evaluate Mixture cure model fitted by rcure}
\usage{
eva_rcure(fit, model)
}
\arguments{
\item{fit}{an extended smcure object}

\item{model}{the type of survival model ("PH", "PO","Normal")}
}
\description{
Evaluate Mixture cure model fitted by rcure
}

