% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpredictrcure.R
\name{plotpredictrcure}
\alias{plotpredictrcure}
\title{Plots predicted rcure object}
\usage{
plotpredictrcure(object, type = "S", xlab = "Time",
  ylab = "Predicted Survival Probability", model = c("ph", "aft"), ...)
}
\arguments{
\item{object}{an object of the predictrcure function}

\item{type}{type of plot. "S" means steps plot.}

\item{xlab}{label for the x axis}

\item{ylab}{label for the y axis}

\item{model}{specifies survival part in cure model, "ph" or "aft"}

\item{...}{further arguments to be passed to the plotpredictrcure function}
}
\description{
Plots predicted survival curve(s) from the estimated robust cure model
}
\references{
Cai, C., Zou, Y., Peng, Y., & Zhang, J. (2012). smcure: An R-Package for estimating semiparametric mixture cure models. Computer methods and programs in biomedicine, 108(3), 1255-1260
}

