% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictrcure.R
\name{predictrcure}
\alias{predictrcure}
\title{Prediction of robust cure model}
\usage{
predictrcure(object, newX, newZ, model = c("ph", "aft"), ...)
}
\arguments{
\item{object}{an object of rcure}

\item{newX}{new value(s) of X}

\item{newZ}{new value(s) of Z}

\item{model}{specifies survival part in cure model, "ph" or "aft"}

\item{...}{further arguments to be passed to the predictrcure function}
}
\description{
Prediction of survival probability from rcure() in order to pass subjects to plotpredictrcure().
}
\references{
Cai, C., Zou, Y., Peng, Y., & Zhang, J. (2012). smcure: An R-Package for estimating semiparametric mixture cure models. Computer methods and programs in biomedicine, 108(3), 1255-1260
}

